Move extensions on QueueDescription to QueueDescriptionExtensions. Do not pollute QueueDescription with responsibilities that don't belong to it.
