+++ src/Microsoft.Azure.ServiceBus/Management/QueueDescription.cs
@@ -117,12 +117,7 @@ namespace Microsoft.Azure.ServiceBus.Management
         {
             get
             {
-                if (this.authorizationRules == null)
-                {
-                    this.authorizationRules = new AuthorizationRules();
-                }
-
-                return this.authorizationRules;
+                return this.authorizationRules ?? (this.authorizationRules = new AuthorizationRules());
             }
             internal set
             {
@@ -176,185 +171,11 @@ namespace Microsoft.Azure.ServiceBus.Management
 
         public bool EnablePartitioning { get; set; } = false;
 
-        static internal QueueDescription ParseFromContent(string xml)
-        {
-            var xDoc = XElement.Parse(xml);
-            if (!xDoc.IsEmpty)
-            {
-                if (xDoc.Name.LocalName == "entry")
-                {
-                    return ParseFromEntryElement(xDoc);
-                }
-            }
-
-            throw new MessagingEntityNotFoundException("Queue was not found");
-        }
-
-        static internal IList<QueueDescription> ParseCollectionFromContent(string xml)
-        {
-            var xDoc = XElement.Parse(xml);
-            if (!xDoc.IsEmpty)
-            {
-                if (xDoc.Name.LocalName == "feed")
-                {
-                    var queueList = new List<QueueDescription>();
-
-                    var entryList = xDoc.Elements(XName.Get("entry", ManagementClientConstants.AtomNs));
-                    foreach (var entry in entryList)
-                    {
-                        queueList.Add(ParseFromEntryElement(entry));
-                    }
-
-                    return queueList;
-                }
-            }
-
-            throw new MessagingEntityNotFoundException("Queue was not found");
-        }
-
-        // TODO: Revisit all properties and ensure they are populated.
-        static private QueueDescription ParseFromEntryElement(XElement xEntry)
-        {
-            try
-            {
-                var name = xEntry.Element(XName.Get("title", ManagementClientConstants.AtomNs)).Value;
-                var qd = new QueueDescription(name);
-
-                var qdXml = xEntry.Element(XName.Get("content", ManagementClientConstants.AtomNs))?
-                    .Element(XName.Get("QueueDescription", ManagementClientConstants.SbNs));
-
-                if (qdXml == null)
-                {
-                    throw new MessagingEntityNotFoundException("Queue was not found");
-                }
-
-                foreach (var element in qdXml.Elements())
-                {
-                    switch (element.Name.LocalName)
-                    {
-                        case "MaxSizeInMegabytes":
-                            qd.MaxSizeInMB = long.Parse(element.Value);
-                            break;
-                        case "RequiresDuplicateDetection":
-                            qd.RequiresDuplicateDetection = bool.Parse(element.Value);
-                            break;
-                        case "RequiresSession":
-                            qd.RequiresSession = bool.Parse(element.Value);
-                            break;
-                        case "DeadLetteringOnMessageExpiration":
-                            qd.EnableDeadLetteringOnMessageExpiration = bool.Parse(element.Value);
-                            break;
-                        case "DuplicateDetectionHistoryTimeWindow":
-                            qd.DuplicateDetectionHistoryTimeWindow = XmlConvert.ToTimeSpan(element.Value);
-                            break;
-                        case "LockDuration":
-                            qd.LockDuration = XmlConvert.ToTimeSpan(element.Value);
-                            break;
-                        case "DefaultMessageTimeToLive":
-                            qd.DefaultMessageTimeToLive = XmlConvert.ToTimeSpan(element.Value);
-                            break;
-                        case "MaxDeliveryCount":
-                            qd.MaxDeliveryCount = int.Parse(element.Value);
-                            break;
-                        case "EnableBatchedOperations":
-                            qd.EnableBatchedOperations = bool.Parse(element.Value);
-                            break;
-                        case "Status":
-                            qd.Status = (EntityStatus)Enum.Parse(typeof(EntityStatus), element.Value);
-                            break;
-                        case "AutoDeleteOnIdle":
-                            qd.AutoDeleteOnIdle = XmlConvert.ToTimeSpan(element.Value);
-                            break;
-                        case "EnablePartitioning":
-                            qd.EnablePartitioning = bool.Parse(element.Value);
-                            break;
-                        case "ForwardTo":
-                            if (!string.IsNullOrWhiteSpace(element.Value))
-                            {
-                                qd.ForwardTo = element.Value;
-                            }
-                            break;
-                        case "ForwardDeadLetteredMessagesTo":
-                            if (!string.IsNullOrWhiteSpace(element.Value))
-                            {
-                                qd.ForwardDeadLetteredMessagesTo = element.Value;
-                            }
-                            break;
-                        case "AuthorizationRules":
-                            qd.AuthorizationRules = AuthorizationRules.ParseFromXElement(element);
-                            break;
-                    }
-                }
-
-                return qd;
-            }
-            catch (Exception ex) when (!(ex is ServiceBusException))
-            {
-                throw new ServiceBusException(false, ex);
-            }
-        }
-
-        internal void NormalizeDescription(string baseAddress)
-        {
-            if (!string.IsNullOrWhiteSpace(this.ForwardTo))
-            {
-                this.ForwardTo = NormalizeForwardToAddress(this.ForwardTo, baseAddress);
-            }
-
-            if (!string.IsNullOrWhiteSpace(this.ForwardDeadLetteredMessagesTo))
-            {
-                this.ForwardDeadLetteredMessagesTo = NormalizeForwardToAddress(this.ForwardDeadLetteredMessagesTo, baseAddress);
-            }
-        }
-
-        private static string NormalizeForwardToAddress(string forwardTo, string baseAddress)
-        {
-            Uri forwardToUri;
-            if (!Uri.TryCreate(forwardTo, UriKind.Absolute, out forwardToUri))
-            {
-                if (!baseAddress.EndsWith("/", StringComparison.Ordinal))
-                {
-                    baseAddress += "/";
-                }
-
-                forwardToUri = new Uri(new Uri(baseAddress), forwardTo);
-            }
-
-            return forwardToUri.AbsoluteUri;
-        }
-
-        internal XDocument Serialize()
-        {
-            XDocument doc = new XDocument(
-                new XElement(XName.Get("entry", ManagementClientConstants.AtomNs),
-                    new XElement(XName.Get("content", ManagementClientConstants.AtomNs),
-                        new XAttribute("type", "application/xml"),
-                        new XElement(XName.Get("QueueDescription",ManagementClientConstants.SbNs),
-                            new XElement(XName.Get("LockDuration", ManagementClientConstants.SbNs), XmlConvert.ToString(this.LockDuration)),
-                            new XElement(XName.Get("MaxSizeInMegabytes", ManagementClientConstants.SbNs), XmlConvert.ToString(this.MaxSizeInMB)),
-                            new XElement(XName.Get("RequiresDuplicateDetection", ManagementClientConstants.SbNs), XmlConvert.ToString(this.RequiresDuplicateDetection)),
-                            new XElement(XName.Get("RequiresSession", ManagementClientConstants.SbNs), XmlConvert.ToString(this.RequiresSession)),
-                            this.DefaultMessageTimeToLive != TimeSpan.MaxValue ? new XElement(XName.Get("DefaultMessageTimeToLive", ManagementClientConstants.SbNs), XmlConvert.ToString(this.DefaultMessageTimeToLive)) : null,
-                            new XElement(XName.Get("DeadLetteringOnMessageExpiration", ManagementClientConstants.SbNs), XmlConvert.ToString(this.EnableDeadLetteringOnMessageExpiration)),
-                            this.RequiresDuplicateDetection && this.DuplicateDetectionHistoryTimeWindow != default ?
-                                new XElement(XName.Get("DuplicateDetectionHistoryTimeWindow", ManagementClientConstants.SbNs), XmlConvert.ToString(this.DuplicateDetectionHistoryTimeWindow))
-                                : null,
-                            new XElement(XName.Get("MaxDeliveryCount", ManagementClientConstants.SbNs), XmlConvert.ToString(this.MaxDeliveryCount)),
-                            new XElement(XName.Get("EnableBatchedOperations", ManagementClientConstants.SbNs), XmlConvert.ToString(this.EnableBatchedOperations)),
-                            this.authorizationRules != null ? this.AuthorizationRules.Serialize() : null,
-                            new XElement(XName.Get("Status", ManagementClientConstants.SbNs), this.Status.ToString()),
-                            this.ForwardTo != null ? new XElement(XName.Get("ForwardTo", ManagementClientConstants.SbNs), this.ForwardTo) : null,
-                            this.AutoDeleteOnIdle != TimeSpan.MaxValue ? new XElement(XName.Get("AutoDeleteOnIdle", ManagementClientConstants.SbNs), XmlConvert.ToString(this.AutoDeleteOnIdle)) : null,
-                            new XElement(XName.Get("EnablePartitioning", ManagementClientConstants.SbNs), XmlConvert.ToString(this.EnablePartitioning)),
-                            this.ForwardDeadLetteredMessagesTo != null ? new XElement(XName.Get("ForwardDeadLetteredMessagesTo", ManagementClientConstants.SbNs), this.ForwardDeadLetteredMessagesTo) : null
-                        ))
-                    ));
-
-            return doc;
-        }
 
         public bool Equals(QueueDescription other)
         {
+            // TODO: other could be null
+
             if (this.Path.Equals(other.Path, StringComparison.OrdinalIgnoreCase)
                 && this.AutoDeleteOnIdle.Equals(other.AutoDeleteOnIdle)
                 && this.DefaultMessageTimeToLive.Equals(other.DefaultMessageTimeToLive)
